# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: PEPA — Platform Engineering & Pipeline Automator API
  description: |
    PEPA — Platform Engineering & Pipeline Automator API for managing services, workflows, scorecards,
    plugins, RBAC, and platform operations.
  version: "0.1.0"
  contact:
    name: PEPA Team
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8088/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /entities:
    get:
      tags: [Entities]
      summary: List entities
      operationId: listEntities
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 20 }
        - name: type_key
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated entity list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Entity"
                  total: { type: integer }
                  page: { type: integer }
                  per_page: { type: integer }
                  total_pages: { type: integer }
    post:
      tags: [Entities]
      summary: Create entity
      operationId: createEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntityRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
  /entities/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Entities]
      summary: Get entity by ID
      operationId: getEntity
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
    put:
      tags: [Entities]
      summary: Update entity
      operationId: updateEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entity"
    delete:
      tags: [Entities]
      summary: Delete entity
      operationId: deleteEntity
      responses:
        "200":
          description: Deleted
  /entities/{id}/graph:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Relationships]
      summary: Get entity relationship graph
      operationId: getEntityGraph
      parameters:
        - name: depth
          in: query
          schema: { type: integer, default: 2 }
      responses:
        "200":
          description: Graph with nodes and edges
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: "#/components/schemas/GraphNode"
                  edges:
                    type: array
                    items:
                      $ref: "#/components/schemas/GraphEdge"
  /entities/{id}/relationships:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Relationships]
      summary: Get entity relationships
      operationId: getEntityRelationships
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  relationships:
                    type: array
                    items:
                      $ref: "#/components/schemas/Relationship"
    post:
      tags: [Relationships]
      summary: Create relationship
      operationId: createRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_id, type_key]
              properties:
                target_id: { type: string, format: uuid }
                type_key: { type: string }
                metadata: { type: object }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relationship"
  /entities/relationships/{relId}:
    parameters:
      - name: relId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Relationships]
      summary: Delete relationship
      operationId: deleteRelationship
      responses:
        "200":
          description: Deleted

  # ── Entity Types ────────────────────────────────────────────────
  /entity-types:
    get:
      tags: [Entity Types]
      summary: List entity types
      operationId: listEntityTypes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  entity_types:
                    type: array
                    items:
                      $ref: "#/components/schemas/EntityType"
    post:
      tags: [Entity Types]
      summary: Register entity type
      operationId: registerEntityType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type_key, display_name]
              properties:
                type_key: { type: string }
                display_name: { type: string }
                description: { type: string }
                category: { type: string }
                icon: { type: string }
      responses:
        "201":
          description: Created

  # ── Workflows ───────────────────────────────────────────────────
...
  /scorecards/entity/{entityId}:
    parameters:
      - name: entityId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Scorecards]
      summary: Get all scorecard results for an entity
      operationId: getEntityScores
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  scores:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScorecardResult"
                  total: { type: integer }

  # ── Audit ───────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Entity:
      type: object
      properties:
        id: { type: string, format: uuid }
        type_key: { type: string }
        name: { type: string }
        description: { type: string }
        status: { type: string }
        metadata: { type: object }
        tenant_id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        sync_status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateEntityRequest:
      type: object
      required: [name, type_key]
      properties:
        name: { type: string }
        type_key: { type: string }
        description: { type: string }
        metadata: { type: object }
    UpdateEntityRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        status: { type: string }
        metadata: { type: object }
    EntityType:
      type: object
      properties:
        id: { type: string, format: uuid }
        type_key: { type: string }
        display_name: { type: string }
        description: { type: string }
        category: { type: string }
        icon: { type: string }
        is_system: { type: boolean }
    Relationship:
      type: object
      properties:
        id: { type: string, format: uuid }
        type_key: { type: string }
        source_id: { type: string, format: uuid }
        target_id: { type: string, format: uuid }
        metadata: { type: object }
        created_at: { type: string, format: date-time }
    GraphNode:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        type_key: { type: string }
        status: { type: string }
    GraphEdge:
      type: object
      properties:
        id: { type: string, format: uuid }
        source: { type: string, format: uuid }
        target: { type: string, format: uuid }
        type_key: { type: string }
...
    ScorecardResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        scorecard_id: { type: string, format: uuid }
        entity_id: { type: string, format: uuid }
        score: { type: integer }
        max_score: { type: integer }
        pass_count: { type: integer }
        fail_count: { type: integer }
        total_rules: { type: integer }
        level: { type: string }
        details: { type: array, items: { type: object } }
        evaluated_at: { type: string, format: date-time }
...